split out tests to separate module
